+++ gridwxcomp/__init__.py
@@ -4,7 +4,7 @@ A package for comparing climate station time series data to `gridMET <http://www
 
 __name__ = 'gridwxcomp'
 __author__ = 'John Volk and Chris Pearson'
-__version__ = '0.0.57'
+__version__ = '0.0.65'
 
 
 from gridwxcomp.prep_input import prep_input
@@ -12,5 +12,3 @@ from gridwxcomp.download_gridmet_ee import download_gridmet_ee
 from gridwxcomp.calc_bias_ratios import calc_bias_ratios
 from gridwxcomp.interpgdal import InterpGdal
 from gridwxcomp.spatial import make_points_file, make_grid, interpolate
-from gridwxcomp.daily_comparison import daily_comparison
-from gridwxcomp.monthly_comparison import monthly_comparison